--- config/evaluation/finetuning.yaml
@@ -1,13 +1,0 @@
-# @package _global_
-evaluation_name: finetuning
-
-evaluation_module:
-  _target_: models.evaluation.Finetuning
-  learning_rate: ${best_learning_rate.finetuning}
-  optimizer: adam
-  # to record both top-1 and top-5 set:[1, 5]
-  top_k: [1]
-  track_per_class_accuracy: False
-
-trainer:
-  max_steps: 5000
