--- config/evaluation/linear.yaml
@@ -1,14 +1,0 @@
-# @package _global_
-
-evaluation_name: linear_eval
-
-evaluation_module:
-  _target_: models.evaluation.LinearEval
-  learning_rate: ${best_learning_rate.linear_eval}
-  optimizer: adam
-  # to record both top-1 and top-5 set:[1, 5]
-  top_k: [1]
-  track_per_class_accuracy: False
-
-trainer:
-  max_steps: 5000
