--- config/evaluation/store_predictions.yaml
@@ -1,11 +1,0 @@
-# @package _global_
-
-evaluation_name: store_predictions
-
-evaluation_module:
-  _target_: models.evaluation.StorePredictions
-  eval_type: finetuning
-  # ckpt_dir: '/checkpoint/marksibrahim/logs/robustness-limits/vit_pretrained_1k_shapenet_random_class_split/2022-09-02_14-20-36/58'
-  # for tmp use /checkpoint/marksibrahim/tmp/robustness-limits/predictions
-  save_dir: '/checkpoint/marksibrahim/results/robustness-limits/predictions'
-
