+++ config/property_library/all.yaml
@@ -3,10 +3,10 @@ property_libary_name: all
 
 ###### Equivariance Measures ######
 equivariance_aug: 
-  _target_: properties.equivariance.Aug_Approximation
-  logging_prefix: 'equivariance_aug'
+  _target_: properties.equivariance.AugmentationEquivariance
+  logging_name: 'equivariance_aug'
  
 ###### Disentanglement Measures ######
 disentanglement_dci: 
   _target_: properties.disentanglement.DCI
-  logging_prefix: 'disentanglement_dci'
+  logging_name: 'disentanglement_dci'
